Use viewBox attribute over size defined on iconset
If a viewBox attribute exists on the source svg
element then use it, otherwise fall back to the size
defined on the iconset.

This allows icons to be shapes other than square, and
allows each icon within a given set to define it's own
viewBox size.

When the viewBox attribute is used, sizing/scaling icons
relative to one another is the responsibility of the user.