Addresses a perf regression related to mirroring
The original implementation computed direction specific to the subtree
where the icon is to be stamped. This is the most correct approach, but
also incurs some non-trivial cost that can add up quickly
(getComputedStyle must be called for every mirrored icon).

The implementation has been changed so that direction is only computed
once per page load per iconset when a mirrored icon is stamped. This
should reduce the number of getComputedStyle calls to 1 in most
scenarios.